requirements.txt
Django==2.1.*
django-ipware==2.1.0
docutils==0.14
future==0.16.0
Pillow==6.1.0
plotly==4.0.0
requests==2.20.1
retrying==1.3.3
semantic-version==2.5.0
sgmllib3k==1.0.0
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.7
